#224. Allow to expand barline slices
When a system is formed by empty measures is not possible to justify the
system as they are fixed width and no force will expand them. This
problem is fixed by:
- making the barline slice expandable (ci > 0 && fi < infinite)
- right aligning the barline shape (is moved to be on xRi).
As the slice is expandible, all extra space is placed before xRi. And as
the shape is right aligned, we ensure that this extra space is placed
just before the barline.

This change also removes the constrain of not transferring space to
barlines when joining the slices, although the code for this has not
been changed in this commit.