#226. Implement auto-scaling when the paper height is too small
A document content scaling factor has been defined. It is used for
scaling the content of the document pages. Paper and paper margins are
not affected, only the content inside the page matgins. This is equivalent
to scaling paper size in the opposite direction, e.g.:
make the score 10% smaller ==> make paper and page margings 10% greater

This scaling factor is now used in all methods for building the graphic
model, whenever the paper size and page margins are retrieved.

Layouters have been modified for adjusting the scaling factor when the
system height is greater than the page height. This can only occur in
two cases:
1) malformed scores: paper too small for the number of instruments
   and staff size, or
3) importing scores from other formats (e.g.: MusicXML scores without
   defining the paper size).