Fix intermediate shapes not created for spanner RelObjs (#274)
This PR fixes #273. Also fixes the same problem in all current spanner RelObjs (wedges, volta brackets and octave shifts).
Changes

Method create_intermediate_shape() is protected in engravers. The public method that triggers its invocation is create_first_or_intermediate_shape(). For spanner RelObjs, the engraver invokes create_intermediate_shape() to create a shape for the whole system when the start point is in a previous sytem and the end point is in a later system.

The problem causing that method create_first_or_intermediate_shape() was not invoked for intermediate shapes was that pending objects lists m_notFinishedRelObj and m_notFinishedLyrics were declared as member variables in SystemLayouter and thus, they were erased after finishing engraving each system. Now its scope is ScoreLayouter scope.

Once this was solved, and create_intermediate_shape() is invoked, the engraver has no information about current system: position, margins, etc. So it was necessary to modify create_first_or_intermediate_shape() signature to pass system information to the engraver.

Then, some small changes in the engravers to fix the position of the intermediate shapes. For octave shifts its create_intermediate_shape() method was empty and was coded.

Finally, struct PendingAuxObjs defined in score layouter was renamed in singular, as PendingAuxObj.